move setup stuff into its own file
